refactor: 프론트 컨트롤러 패턴 적용
